pyproject.toml
[tool.poetry]
name = "haystack"
version = "0.1.0"
description = "haystack test"
authors = ["Zachary Royals"]

[tool.poetry.dependencies]
python = "^3.9"
requests-html = "^0.10.0"
fastapi = "^0.70.0"
sodapy = "^2.1.0"
pandas = "^1.3.4"
tor = "^1.0.0"
selenium = "^4.0.0"
stem = "^1.8.0"
bs4 = "^0.0.1"
s3fs = "^2021.10.1"
fastparquet = "^0.7.1"
prefect = {extras = ["viz"], version = "^0.15.7"}
jupyter = "^1.0.0"
python-dotenv = "^0.19.1"
graphviz = "^0.17"
dask = "^2021.10.0"
autopep8 = "^1.5.7"
DateTime = "^4.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
